# api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v2/users/me/onboarding": {
      "post": {
        "operationId": "UsersController_completeOnboarding",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDetailsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Onboarded user profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Complete user onboarding",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v2/users/me": {
      "get": {
        "operationId": "UsersController_findMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get current user",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update current user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_removeMe",
        "parameters": [
          {
            "name": "photos",
            "required": true,
            "in": "query",
            "description": "What happens to photos the account uploaded into events that outlive it. KEEP: they stay in the event with no uploader. DELETE: they are removed everywhere. Uploads still in progress are always discarded. The choice is stored with the deletion intent, so a resumed saga honours it. Required: the two outcomes are both irreversible, so a caller that omits it gets a 400 rather than a guess.",
            "schema": {
              "type": "string",
              "enum": [
                "KEEP",
                "DELETE"
              ]
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted (empty data envelope at runtime)"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete current user",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v2/users/me/storage": {
      "get": {
        "operationId": "UsersController_getMyStorage",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Photo storage usage",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserStorageResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get current user photo storage quota",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v2/users/me/avatar/upload-url": {
      "post": {
        "operationId": "UsersController_createAvatarUploadUrl",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateImageUploadDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload id with a presigned S3 PUT URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ImageUploadResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Mint a presigned upload URL for the current user's avatar",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v2/users/me/avatar": {
      "put": {
        "operationId": "UsersController_confirmAvatarUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmImageUploadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User profile with the new avatar",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Confirm an uploaded avatar and set it on the current user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_removeAvatar",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Avatar removed (empty data envelope at runtime)"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Remove the current user's avatar",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Everglow API",
    "description": "Photo-sharing platform for events — HTTP API contract (OpenAPI 3)",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "JWT access token"
      }
    },
    "schemas": {
      "UserDetailsResponseDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true,
            "maxLength": 2048,
            "description": "Short-lived presigned URL of the profile avatar; null when none is set"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "email",
          "name",
          "avatarUrl",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "isOnboarded": {
            "type": "boolean"
          },
          "details": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/UserDetailsResponseDto"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "isOnboarded",
          "details",
          "createdAt",
          "updatedAt"
        ]
      },
      "ResponseMetaDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "example": "2026-06-03T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v2/auth/signin"
          }
        },
        "required": [
          "timestamp",
          "path"
        ]
      },
      "CreateUserDetailsDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe",
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "maxLength": 255
          }
        },
        "required": [
          "name",
          "email"
        ]
      },
      "UserStorageResponseDto": {
        "type": "object",
        "properties": {
          "usedBytes": {
            "type": "string",
            "description": "Bytes currently used by the caller's uploads",
            "example": "2147483648"
          },
          "limitBytes": {
            "type": "string",
            "description": "Maximum bytes the caller may use",
            "example": "5368709120"
          },
          "remainingBytes": {
            "type": "string",
            "description": "Bytes remaining before the quota is reached",
            "example": "3221225472"
          }
        },
        "required": [
          "usedBytes",
          "limitBytes",
          "remainingBytes"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe",
            "maxLength": 255
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "maxLength": 255
          }
        }
      },
      "ImageUploadResponseDto": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "format": "uuid",
            "description": "Pass it back to confirm the upload"
          },
          "uploadUrl": {
            "type": "string",
            "maxLength": 2048,
            "description": "Presigned S3 PUT URL the client uploads bytes to"
          }
        },
        "required": [
          "uploadId",
          "uploadUrl"
        ]
      },
      "CreateImageUploadDto": {
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ],
            "example": "image/jpeg"
          },
          "sizeBytes": {
            "type": "number",
            "minimum": 1,
            "maximum": 5242880,
            "example": 262144
          }
        },
        "required": [
          "contentType",
          "sizeBytes"
        ]
      },
      "ConfirmImageUploadDto": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "format": "uuid",
            "description": "The uploadId returned when the upload URL was minted"
          }
        },
        "required": [
          "uploadId"
        ]
      },
...
    },
    "responses": {
      "TooManyRequests": {
        "description": "Rate limit exceeded; retry after the number of seconds in the Retry-After header",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "meta"
              ],
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Too many requests, please try again later"
                },
                "code": {
                  "type": "string",
                  "description": "Stable machine-readable error code, when the error has one",
                  "example": "RATE_LIMIT_EXCEEDED"
                },
                "meta": {
                  "$ref": "#/components/schemas/ResponseMetaDto"
                }
              }
            }
          }
        }
      }
    }
  }
}